+++ sdk/python/kfp/containers/_gcs_helper.py
